backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic-settings==2.6.0
uuid7==0.1.0
python-dotenv==1.0.1
pytest==8.3.3
httpx==0.27.2
passlib[bcrypt]==1.7.4
pyjwt==2.9.0
python-multipart==0.0.12